Refactor EVM chain ID retrieval by replacing the coin type extraction with a new utility function, getTwChainId, for improved clarity and maintainability. Additionally, introduce a new utility function, getKeysignTwPublicKey, to streamline public key handling in transaction input data for Ripple and Ton chains.
