+++ core/mpc/keysign/tw/getTwChainId.ts
@@ -1,0 +1,21 @@
+import { Chain } from '@core/chain/Chain'
+import { getCoinType } from '@core/chain/coin/coinType'
+import { WalletCore } from '@trustwallet/wallet-core'
+
+type Input = {
+  walletCore: WalletCore
+  chain: Chain
+}
+
+export const getTwChainId = ({ walletCore, chain }: Input) => {
+  if (chain === Chain.MayaChain) {
+    return 'mayachain-mainnet-v1'
+  }
+
+  const coinType = getCoinType({
+    walletCore,
+    chain,
+  })
+
+  return walletCore.CoinTypeExt.chainId(coinType)
+}
