+++ core/mpc/keysign/txInputData/ripple.ts
@@ -3,7 +3,7 @@ import { TW } from '@trustwallet/wallet-core'
 import Long from 'long'
 
 import { getBlockchainSpecificValue } from '../chainSpecific/KeysignChainSpecific'
-import { getKeysignTwPublicKey } from '../utils/getKeysignTwPublicKey'
+import { getKeysignTwPublicKey } from '../tw/getKeysignTwPublicKey'
 import { TxInputDataResolver } from './TxInputDataResolver'
 
 export const getRippleTxInputData: TxInputDataResolver<'ripple'> = ({
